requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.27.0
google-generativeai==0.3.2
langchain==0.1.0
langchain-google-genai==0.0.9
celery==5.3.6
python-dotenv==1.0.1
# Testing dependencies
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
# redis is intentionally omitted as per the 'no redis' constraint